backend/src/requirements.txt
astroid==2.2.5
autopep8==1.4.3
beautifulsoup4==4.7.1
colorama==0.4.1
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
django-tinymce4-lite==1.7.4
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
isort==4.3.16
jsmin==2.2.2
lazy-object-proxy==1.3.1
lxml==4.3.1
mccabe==0.6.1
numpy==1.16.2
pandas==0.24.1
pep8==1.7.1
postmen==1.2
psutil==5.5.1
psycopg2==2.7.7
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.7.2
pypiwin32==223
python-dateutil==2.8.0
pytz==2018.9
pywin32==224
requests==2.11.1
selenium==3.141.0
six==1.12.0
soupsieve==1.7.3
tornado==5.1.1
typed-ast==1.3.1
urllib3==1.24.1
WMI==1.4.9
wrapt==1.11.1
